Added more path to the prompt.
What
===
Added more path to the prompt. Up to 50 characters will be shown now
instead of 30, and it will display as much of the path as possible
rather than only show the latest directory.

Why
===
Now that the prompt is on a separate line, there is more room for this.